+++ catbuffer/parser/CHANGELOG.md
@@ -3,12 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0] - 19-Apr-2021
+## [next]
 
 ### Added
- - initial code release
+ - YAML export option
 
-## [2.0.0] - 30-Aug-2021
+### Changed
+ - parser output to YAML instead of python objects
+
+## [2.0.0] - 02-Sep-2021
 
 ### Added
  - first class support for reserved fields with 'make_reserved' keyword
@@ -21,6 +24,11 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
  - replace `const` keyword with `make_const` to preserve a unified left to right flow
  - rename keyword 'has' to 'in'
  - deprecate explicit use of 'byte' type
+
+## [1.0.0] - 19-Apr-2021
+
+### Added
+ - initial code release
 
 [2.0.0]: https://github.com/symbol/catbuffer-parser/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/symbol/catbuffer-parser/releases/tag/v1.0.0